+++ CHANGELOG.md
@@ -1,5 +1,20 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r231-release...ffe8f9fb5ac504c8752d0483f056ff1dea02f2cf)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r232-release...d6d3625807e91b24fd4a4c939a33d481dc57e881)
 
-- Update Cell.GetUnitFrame
-- Delay creation of some indicators
-- Fix a critical issue
+- Add "Blizzard 4" for Role Icon indicator
+- Add desc for AoE Healing
+- Add "current+" for Dispel Highlight
+- Fix click-casting for "Rescue"
+- Fix layout preview
+- Fix wrath buff tracker
+- Fix raid debuff creation
+- Fix frame level of Actions
+- Fix "Invert Color"
+- Fix "Track by name"
+- Fix block color options
+- Fix PWS id
+- Update range check
+- Update locales
+- Update built-in spells list button alpha
+- Update frame levels
+- Lower aggro level
+- GetSpellNameAndIcon -> GetSpellInfo